+++ modules/canbus/vehicle/ch/protocol/steer_command_112.cc
@@ -62,7 +62,7 @@ void Steercommand112::set_p_steer_angle_en_ctrl(
   int x = steer_angle_en_ctrl;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 0, 8);
+  to_set.set_value(static_cast<uint8_t>(x), 0, 8);
 }
 
 Steercommand112* Steercommand112::set_steer_angle_cmd(double steer_angle_cmd) {
@@ -77,15 +77,15 @@ Steercommand112* Steercommand112::set_steer_angle_cmd(double steer_angle_cmd) {
 void Steercommand112::set_p_steer_angle_cmd(uint8_t* data,
                                             double steer_angle_cmd) {
   steer_angle_cmd = ProtocolData::BoundedValue(-0.524, 0.524, steer_angle_cmd);
-  int x = steer_angle_cmd / 0.001000;
+  int x = static_cast<int>(steer_angle_cmd / 0.001000);
   uint8_t t = 0;
 
-  t = x & 0xFF;
+  t = static_cast<uint8_t>(x & 0xFF);
   Byte to_set0(data + 1);
   to_set0.set_value(t, 0, 8);
   x >>= 8;
 
-  t = x & 0xFF;
+  t = static_cast<uint8_t>(x & 0xFF);
   Byte to_set1(data + 2);
   to_set1.set_value(t, 0, 8);
 }
